==> 02_model_selection.Rmd <==
---
title: "Model selection"
date: "`r Sys.Date()`"
output: 
  html_document:
    number_sections: true
    toc: true
    toc_depth: 3
    toc_float: true
---
<style>
.main-container {
  max-width: 1400px !important;  /* Adjust the width as needed */
}

pre, code {
  white-space: pre-wrap;  /* Ensures that long lines wrap */
  word-wrap: break-word;  /* Breaks long words if necessary */
}

pre {
  max-width: 100% !important;
  width: 100% !important;
  overflow-x: auto !important; /* Add horizontal scrollbar if content overflows */
  white-space: pre-line !important; /* Convert line breaks to spaces */
  word-wrap: break-word !important; /* Break long words if necessary */
  word-break: break-all !important; /* Break words to fit the container */
  font-size: 14px !important; /* Adjust font size if needed */
}

</style>


The purpose of this document is to present an abridged version of our model selection process.

```{r setup, message = FALSE, eval=TRUE}
knitr::opts_chunk$set(
	message = FALSE,
	warning = TRUE,
	include = TRUE,
	error = TRUE,
	fig.width = 8,
	fig.height = 4
)

library(tidyverse)
library(haven)
library(ggrepel)
library(brms)
library(tidybayes)
library(marginaleffects)
library(bayesplot)
library(rstan)

# depending on the platform on which you want to run the brm you might need this or not. We ran the models on a Linux-operated server, cmdstanr version 0.5.3

options(mc.cores = 4,
        brms.backend = "cmdstanr",
        scipen = 999,
        digits = 4,
        width = 140)

# windowsFonts(Times = windowsFont("Times New Roman"))
theme_set(theme_minimal(base_size = 12, base_family = "Times"))

# get data
load("data/preprocessed/de_census/census.Rda")
load("../unshareable_data/preprocessed/tl.Rda")

# get function that calculates norms and plots them
source("age_norm_comparisons.R")
```




# Model comparisons
## Normal likelihood
### No interactions

We start with a main effect only model with age as a spline for regularised continuous modelling, and sex as a fixed instead of random effect because of the small number of categories (2).


```{r}
brm_nor <-
  brm(bf(
      cft   ~ (1 | mig) + (1 | educ) + male + s(age),
      sigma ~ (1 | mig) + (1 | educ) + male + s(age)
      ),
      family = gaussian(),
      chains = 4,
      seed = 810,
      file = "../unshareable_data/brms/cft/brm_nor",
      data = tl) %>% 
  add_criterion("loo")

brm_nor


pp_check(brm_nor)

bayes_R2(brm_nor)
```

Doesn't look too bad but we can do better, e.g., by fitting a spline for each education category (in line with an interaction between age and educational attainment).


### Spline by educ for the mean
Now we fit a spline for each education category separately to allow for different development trajectories. 

#### educ as random main effect
```{r}


brm_nor_s_by_educ_main_random_educ <-
  brm(bf(
      cft ~ (1 | mig) + male + (1 | educ) + s(age, by = educ),
      sigma ~ (1 | mig) + (1 | educ) + male + s(age)
      ),
      family = gaussian(),
      chains = 4,
      seed = 810,
      file = "../unshareable_data/brms/cft/brm_nor_s_by_educ_main_random_educ",
      data = tl) %>% 
  add_criterion("loo")


brm_nor_s_by_educ_main_random_educ


pp_check(brm_nor_s_by_educ_main_random_educ)

loo_compare(brm_nor, brm_nor_s_by_educ_main_random_educ)


```
179 divergent transitions and negligible gain in fit compared to the simpler model. 

#### educ as fixed main effect
```{r}

brm_nor_age_by_educ_main_fixed_educ <-
  brm(bf(
      cft ~ (1 | mig) + male + educ + s(age, by = educ),
      sigma ~ (1 | mig) + (1 | educ) + male + s(age)
      ),
      family = gaussian(),
      chains = 4,
      seed = 810,
      file = "../unshareable_data/brms/cft/brm_nor_s_by_educ_main_fixed_educ",
      data = tl) %>% 
  add_criterion("loo")

brm_nor_age_by_educ_main_fixed_educ


pp_check(brm_nor_age_by_educ_main_fixed_educ)
```

138 divergent transitions when educ is a fixed effect instead of random.



#### No educ main effect
```{r}
brm_nor_s_by_educ <-
  brm(bf(
      cft ~ (1 | mig) + male + s(age, by = educ),
      sigma ~ (1 | mig) + (1 | educ) + male + s(age)),
      family = gaussian(),
      chains = 4,
      seed = 810,
      file = "../unshareable_data/brms/cft/brm_nor_s_by_educ",
      data = tl) %>% 
  add_criterion("loo")



brm_nor_s_by_educ


pp_check(brm_nor_s_by_educ)


loo_compare(brm_nor, brm_nor_s_by_educ, brm_nor_s_by_educ_main_random_educ, brm_nor_age_by_educ_main_fixed_educ)
```

Somewhat worse prediction accuracy, but similar fit and 82 divergent transitions. 

Here we don't include a main effect for educ because it does not contribute any improvement in fit while leading to two problems: 1. more divergent transitions and 2. irregualrities in the norms, see comparison:

```{r, "educ_vs_not"}

main_educ_vs_not <- age_norm_comparisons(
  brm_nor, brm_nor_s_by_educ, brm_nor_s_by_educ_main_random_educ,
   palette = c(
  "#BC3C29FF",
  # "#0072B5FF",
  "#20854EFF",
  # "#7876B1FF",
  "#6F99ADFF",
  # "#E18727FF",
  # "#FFDC91FF",
  "#EE4C97FF"
),  
  output_file = "data/results/educ_vs_not.rds"
  )

main_educ_vs_not[-1]

```


So it seems the term `s(age, by = educ)` is required without additionally including a main effect of educ to get sensible estimates for ages 19-22. This is probably due to the change in meaning of the education variable at age 19 (school type to ISCED), which causes some difficulties for the model. Note that this issue also afflicts the model with random slopes instead of splines.

#### Random slopes instead of splines?

```{r, "random_slopes_vs_spline"}

brm_nor_random_slopes <-
  brm(bf(
      cft ~ (1 | mig) + male + (1 + age|educ),
      sigma ~ (1 | mig) + male + (1 + age|educ)
      ),
      family = gaussian(),
      chains = 4,
      seed = 810,
      file = "../unshareable_data/brms/cft/brm_nor_educ_random_slope",
      data = tl) %>% 
  add_criterion("loo")

random_slopes_vs_spline <- age_norm_comparisons(
  brm_nor_random_slopes, brm_nor_s_by_educ, 
   palette = c(
  "#BC3C29FF",
  # "#0072B5FF",
  "#20854EFF",
  # "#7876B1FF",
  "#6F99ADFF"
  # "#E18727FF",
  # "#FFDC91FF",
  # "#EE4C97FF"
),  
  output_file = "data/results/random_slopes_vs_spline.rds"
  )

random_slopes_vs_spline[-1]
```


### More interactions
We incorporate more interactions into the mean prediction for theoretical reasons (it is e.g., reasonable to expect educational background to interact with migration background). We do not incorporate interaction terms into the sigma prediction as they would unnecessarily complicate the model for little gain in prediction accuracy.

```{r}

brm_nor_ints <-
  brm(bf(
      cft   ~ (1 | mig) + male + (1 | mig:male) + (1 | mig:educ) + (1 | male:educ)  + s(age, by = educ),
      sigma ~ (1 | mig) + (1 | educ) + male + s(age)
      ),
      chains = 4,
      seed = 810,
      file = "../unshareable_data/brms/cft/brm_nor_ints",
      data = tl) %>% 
  add_criterion("loo")



brm_nor_ints

pp_check(brm_nor_ints)


bayes_R2(brm_nor_ints)

loo_compare(brm_nor, brm_nor_s_by_educ, brm_nor_ints)
```




Somewhat better fit,  still not great + 63 divergent transitions.

How does the RPP look?

```{r, "rpp_nor_ints"}

rpp_nor_ints <- age_norm_comparisons(
  brm_nor_ints,
   palette = c(
  "#BC3C29FF",
  # "#0072B5FF",
  "#20854EFF"
  # "#7876B1FF",
  # "#6F99ADFF",
  # "#E18727FF",
  # "#FFDC91FF",
  # "#EE4C97FF"
),  
  output_file = "data/results/rpp_nor_ints.rds"
  )

rpp_nor_ints[-1]

```


Including the educ:male interaction reintroduces the malignant dip at age 19, so dropping that term in the next model.



### No male:educ interaction

```{r}
brm_nor_ints_no_educ_male <-
  brm(bf(
      cft   ~ (1 | mig) + male + (1 | mig:male) + (1 | mig:educ) + s(age, by = educ),
      sigma ~ (1 | mig) + (1 | educ) + male + s(age)
      ),
      chains = 4,
      seed = 810,
      file = "../unshareable_data/brms/cft/brm_nor_ints_no_educ_male",
      data = tl) %>% 
  add_criterion("loo")



pp_check(brm_nor_ints_no_educ_male)
brm_nor_ints_no_educ_male


loo_compare(brm_nor_ints_no_educ_male, brm_nor_ints)
```




```{r, "educXmale_vs_not"}

educXmale_vs_not <- age_norm_comparisons(
  brm_nor_ints, brm_nor_ints_no_educ_male,
  labels = c("Raw", "All ints", "No Educ*Male Int"),
  prediction_transform = list(
  function(x) round(pmax(0, pmin(56, x))), # for handling normal predictons
  function(x) round(pmax(0, pmin(56, x)))  # for handling normal predictions
),
  palette = c(
  "#BC3C29FF",
  # "#0072B5FF",
  # "#20854EFF",
  # "#7876B1FF",
  "#6F99ADFF",
  # "#E18727FF",
  # "#FFDC91FF",
  "#EE4C97FF"
),  
  output_file = "data/results/educXmale_vs_not.rds"
  )


educXmale_vs_not[-1]
```


Looks much better, this prediction constellation seems okay.




## Alternative likelihood functions

The fit of the Gaussian models seems okay, but it seems that there is certain degree of left skew that the necessarily symmetrical fails to account for. So, here we explore alternative likelihood models that are not necessarily symmetric around their means. 

### Skew normal

A skew normal is like a normal, but parameterised by an additional skew (alpha) parameter that can incorporate non zero skews. One of the chains kept breaking, so we increased the number of chains to 5 to get 4 chains:

```{r}
brm_skew_ints_no_educ_male <-
  brm(bf(
      cft   ~ (1 | mig) + male + (1 | mig:male) + (1 | mig:educ) + s(age, by = educ),
      sigma ~ (1 | mig) + (1 | educ) + male + s(age),
      alpha ~ (1 | mig) + (1 | educ) + male + s(age)
      ),
      family = "skew_normal",
      chains = 5, # 5 chains instead of 4 because 1 chain kept finishing unexpectedly due to bad initial values, so we end up with 4
      seed = 810,
      file = "../unshareable_data/brms/cft/brm_skew_ints_no_educ_male",
      data = tl) %>% 
  add_criterion("loo")

pp_check(brm_skew_ints_no_educ_male)

brm_skew_ints_no_educ_male

loo_compare(brm_skew_ints_no_educ_male, brm_nor_ints_no_educ_male)

```

The skew model has better fit to the data, but skew normal also often has chain initialisation issues so we explore other likelihoods next. 


### Binomial
```{r}
brm_bin_ints_no_educ_male <-
  brm(bf(
      cft | trials(56) ~ (1 | mig) + male + (1 | mig:male) + (1 | mig:educ) + s(age, by = educ)
      ),
      family = binomial(),  
      chains = 4,
      seed = 810,
      file = "../unshareable_data/brms/cft/brm_bin_ints_no_educ_male",
      data = tl) %>% 
  add_criterion("loo")


brm_bin_ints_no_educ_male


pp_check(brm_bin_ints_no_educ_male)




```


Fit pretty bad, probably due to the model's inability to deal with overdispersion.



### Beta

```{r}

brm_beta_ints_no_educ_male <-
  brm(bf(
      cft_prop ~ (1 | mig) + male + (1 | mig:male) + (1 | mig:educ) + s(age, by = educ), # 
      phi ~ (1 | mig) + (1 | educ) + male + s(age)  
      ),
      family = Beta(),  
      chains = 4,
      seed = 810,
      file = "../unshareable_data/brms/cft/brm_beta_ints_no_educ_male",
      data = tl) %>% 
  add_criterion("loo")

brm_beta_ints_no_educ_male


pp_check(brm_beta_ints_no_educ_male)

```

```{r}
ggsave("figures/S05_pp_check.jpeg", width = 8, height = 4)
```

Fit looks good, LOO doesn't make sense because different scales. We're going with this model, rescaling the variable back to original at the poststratification stage.


```{r, "beta_vs_normal"}


beta_vs_normal <- age_norm_comparisons(
  brm_beta_ints_no_educ_male, brm_nor_ints_no_educ_male,
  labels = c( "Raw", "RPP, beta", "RPP, normal"),
  prediction_transform  = list(
  function(x) round(x*56),  # for handling beta proportion predictions
  function(x) round(pmax(0, pmin(56, x))) # for handling normal predictons
),
  palette = c(
  "#BC3C29FF",
  "#0072B5FF",
  # "#20854EFF",
  # "#7876B1FF",
  # "#6F99ADFF",
  # "#E18727FF",
  # "#FFDC91FF",
  "#EE4C97FF" 
), 
  output_file = "data/results/beta_vs_normal.rds"
  )


beta_vs_normal[-1]

```

#### boosted
```{r}

brm_beta_ints_no_educ_male_boosted <-
  brm(bf( # cft_prop is cft sum score divided by 56 (total n of items)
      cft_prop ~ (1 | mig) + male + (1 | mig:male) + (1 | mig:educ) + s(age, by = educ),
      phi ~ (1 | mig) + (1 | educ) + male + s(age)
      ),
      family = Beta(),
      chains = 4,
      iter = 7000,
      control = list(adapt_delta = 0.999),
      seed = 810,
      file = "../unshareable_data/brms/cft/brm_beta_ints_no_educ_male_boosted",
      data = tl) %>% 
  add_criterion("loo")

brm_beta_ints_no_educ_male_boosted

```



```{r, "beta_vs_boosted_beta"}
beta_vs_boosted_beta <- age_norm_comparisons(
  brm_beta_ints_no_educ_male, brm_beta_ints_no_educ_male_boosted,
  labels = c("Raw", "RPP, beta", "RPP, boosted beta"),
  prediction_transform = list(
  function(x) round(x*56),  # for handling beta proportion predictions
  function(x) round(x*56)  # for handling beta proportion predictions
),
   palette = c(
  "#BC3C29FF",
  "#0072B5FF",
  # "#20854EFF",
  # "#7876B1FF",
  # "#6F99ADFF",
  # "#E18727FF",
  # "#FFDC91FF",
  "#EE4C97FF"
),  
  output_file = "data/results/beta_vs_boosted_beta.rds"
  )

beta_vs_boosted_beta[-1]
```

```{r}
ggsave("figures/S04_beta_vs_boosted_beta.jpeg", beta_vs_boosted_beta$means_plot, width = 8, height = 4)
ggsave("figures/S04_beta_vs_boosted_beta_percentile.jpeg", beta_vs_boosted_beta$percentile_plot, width = 8, height = 4)
```



### Beta binomial
Took too long to fit (200 iterations after 40 hours). Setting stronger priors sped up fitting time but led to nonconvergence. 

# Some exploration of the selected model

## Check fit within manual age groups
```{r, fig.height=8, fig.width=8}
t <- tl %>% filter(!is.na(age_group))

yrep <- posterior_predict(brm_beta_ints_no_educ_male, newdata = t, ndraws = 1000)


p <- ppc_pit_ecdf_grouped(t$cft/56, yrep, group = as.factor(t$age_group), prob = 0.99, plot_diff = F)

p

ggsave("figures/S06_ppc_grouped.jpeg", width = 8, height = 8)
```
Good fit within age groups too.




## R2
```{r}
loo_R2(brm_beta_ints_no_educ_male)
```
The modelled predictors explain about 30% of the variance.


## Prediction plots for education and migration background

```{r, fig.height = 5, fig.width = 8}
plot_predictions(brm_beta_ints_no_educ_male,  condition = c("educ"), allow_new_levels = T) + 
  labs(x = "Education", y = "Proportion of correct items") +
theme(axis.text.x = element_text(angle = 60, hjust = 1),
    plot.margin = margin(0.8, 2, 0.8, 2, "cm")) 
```


We see that means differ considerably across categories of `educ`. The large 95% CIs are for the school type categories, i.e., those which have no observations at ages older than 18.

```{r, fig.height = 5, fig.width = 8}
plot_predictions(brm_beta_ints_no_educ_male,  condition = c("mig"), allow_new_levels = T)   + 
  labs(x = "Migration Background", y = "Proportion of correct items") +
theme(axis.text.x = element_text(angle = 60, hjust = 1),
    plot.margin = margin(0.8, 2, 0.8, 1.4, "cm"))
```


Less variation among migration categories.



## Plots of the conditional age splines 

```{r}
distinct_tl <- tl %>%
  select(age, educ, mig, male) %>%
  mutate(male = FALSE, mig = "Citizen: No mig background") %>%
  distinct()

fitted_tl <- fitted(brm_beta_ints_no_educ_male, newdata = distinct_tl, summary = T, ndraws = 1000, probs = c(0.20, 0.80)) %>% as_tibble() %>%
  bind_cols(distinct_tl) %>%
  mutate(
    school_type = case_when(
      str_detect(educ, "ST") ~ educ,
      TRUE ~ NA_character_
    ),
    school_type = str_replace(school_type, "^[^ ]* ", ""),
    isced = case_when(
      str_detect(educ, "ST") ~ NA_character_,
      TRUE ~ educ
    ),
    isced = str_replace(isced, "^[^ ]* ", "")
  )  

fitted_ST <- fitted_tl %>% 
  filter(!is.na(school_type) & age <= 18)

fitted_ISCED <- fitted_tl %>% 
  filter(!is.na(isced) & (age >= 19 & age <= 65) 
         & !(age <= 21 & isced == "5a: Tertiary, e.g., college")
         )



max_age_per_ST <- fitted_ST %>%
                     group_by(school_type) %>%
                     summarise(max_age = max(age, na.rm = TRUE))


labels <- fitted_ST %>%
                inner_join(max_age_per_ST, by = "school_type") %>%
                filter(age == max_age)


# colour palette from https://jacksonlab.agronomy.wisc.edu/2016/05/23/15-level-colorblind-friendly-palette/
ggplot(data = fitted_ST, aes(x = age, y = Estimate, ymin = Q20, ymax = Q80, color = school_type, fill = school_type)) +
  scale_x_continuous(breaks = seq(12, 18, by = 2),
                   expand = expansion(add = c(0, 2))) +
  geom_line(linewidth = 1) +
  geom_ribbon(alpha = 0.2, linewidth = .01) + 
  labs(x = "Age", y = "Proportion of correct items", colour = "School Type") +
  geom_text_repel(data = labels %>% arrange(school_type),
    aes(label = c(school_type[1:5], "", school_type[7])), family = "Times", seed = 810, xlim = c(18.3, 20)) +
  geom_text_repel(data = labels %>% arrange(school_type),
    aes(label = c("", "", "", "", "", school_type[6], "")), family = "Times", nudge_x = 1.2, seed = 810) +
                       # Und bist du erst mein ehlich Weib,
                       # Dann bist du zu beneiden,
                       # Dann lebst du in lauter Zeitvertreib,
                       # In lauter Pläsier und Freuden.
                       
                       # Und wenn du schiltst und wenn du tobst,
                       # Ich werd es geduldig leiden;
                       # Doch wenn du mein ggplot nicht lobst,
                       # Laß ich mich von dir scheiden.
  theme(
    panel.grid.major.y = element_blank(),
    panel.grid.minor.y = element_blank(),
    panel.grid.major.x = element_line(linetype = "dashed", size = 0.3),
    panel.grid.minor.x = element_blank(),
    legend.position = "none"
  ) +
  scale_color_manual(values = c("#009292","#000000","#ff6db6","#006ddb","#920000", "#490092","#6db6ff")) +
  scale_fill_manual(values = c("#009292","#000000","#ff6db6","#006ddb","#920000", "#490092","#6db6ff"))

ggsave("figures/03_splines_ST.jpeg", width = 9, height = 5)
```


```{r}
max_age_per_isced <- fitted_ISCED %>%
                     group_by(isced) %>%
                     summarise(max_age = max(age, na.rm = TRUE))


labels <- fitted_ISCED %>%
                inner_join(max_age_per_isced, by = "isced") %>%
                filter(age == max_age)

ggplot(data = fitted_ISCED, aes(x = age, y = Estimate, ymin = Q20, ymax = Q80, color = isced, fill = isced)) +
  scale_x_continuous(breaks = seq(19, 65, by = 7),
                     expand = expansion(add = c(1, 21)))  +
  geom_line(linewidth = 1) +
  geom_ribbon(alpha = 0.2, linewidth = .01) +
  labs(x = "Age", y = "Proportion of correct itemse", colour = "ISCED 1997 Code")  +
  geom_text_repel(data = labels %>%
           arrange(isced),
    aes(label = isced), family = "Times", seed = 14, xlim = c(65, 100), box.padding = 0.1) + 
  theme(
    panel.grid.major.y = element_blank(),
    panel.grid.minor.y = element_blank(),
    panel.grid.major.x = element_line(linetype = "dashed", size = 0.3),
    panel.grid.minor.x = element_blank(),
    legend.position = "none"
  ) +
  scale_color_manual(values = c("#b66dff","#6db6ff","#009292", "#920000","#924900","#490092","#24ff24","#000000")) +
  scale_fill_manual(values = c("#b66dff","#6db6ff","#009292", "#920000","#924900","#490092","#24ff24","#000000"))


ggsave("figures/03_splines_ISCED.jpeg", width = 9, height = 5, scale = 1)

```

## Distribution of RPP predicitons
### Generate a poststratified sample
```{r, eval=FALSE}
set.seed(14)

sim_pop_sample <- census %>%  
  sample_n(size = 100000,
           weight = census_n,
           replace = TRUE)

sim_pop_sample_with_draws <- sim_pop_sample %>%
  add_predicted_draws(brm_beta_ints_no_educ_male,
                      ndraws = 1000,
                      seed = 810,
                      allow_new_levels = TRUE) %>%
  mutate(.prediction = round(.prediction*56))  %>% 
  mutate(age_group = case_when(
                         age == 11 ~ '11',
                         age == 12 ~ '12',
                         age == 13 ~ '13',
                         age == 14 ~ '14',
                         age == 15 ~ '15',
                         age == 16 ~ '16',
                         age >= 17 & age <= 19 ~ '17-19',
                         age >= 20 & age <= 24 ~ '20-24',
                         age >= 25 & age <= 29 ~ '25-29',
                         age >= 30 & age <= 34 ~ '30-34',
                         age >= 35 & age <= 39 ~ '35-39',
                         age >= 40 & age <= 44 ~ '40-44',
                         age >= 45 & age <= 49 ~ '45-49',
                         age >= 50 & age <= 54 ~ '50-54',
                         age >= 55 & age <= 59 ~ '55-59',
                         age >= 60 & age <= 64 ~ '60-64',
                         TRUE ~ NA_character_))

data.table::fwrite(sim_pop_sample_with_draws, "data/results/sim_pop_sample_with_draws_beta.csv.gz")
```

### Overall distribution
```{r}
# data.table::fread() requires installing the package "R.utils"
sim_pop_sample_with_draws <- data.table::fread("data/results/sim_pop_sample_with_draws_beta.csv.gz")

means_sds_and_ses_RPP <- sim_pop_sample_with_draws %>%
  group_by(age, .draw) %>% 
  summarise(mean_prediction = mean(.prediction), 
            sd_prediction = sd(.prediction)) %>%
  group_by(age) %>% 
  summarise(RPP_mean = mean(mean_prediction), 
            RPP_se_of_mean = sd(mean_prediction), 
            RPP_sd = sqrt(mean(sd_prediction^2)), 
            RPP_se_of_sd = sd(sd_prediction)) 

 sim_pop_sample_with_draws %>% 
  ggplot(aes(x = .prediction)) +
  geom_histogram(binwidth = 1, fill = "darkblue", color = "black") +
  scale_x_continuous(breaks = seq(min(sim_pop_sample_with_draws$.prediction), 
                                   max(sim_pop_sample_with_draws$.prediction), 
                                   by = 4)) +
  theme_minimal() +
  labs(x = "RPP predicted CFT 20-R score", y = "Count")
```





### Distributions witin age groups
```{r, fig.width= 16, fig.height=14}
 sim_pop_sample_with_draws %>% 
  ggplot(aes(x = .prediction)) +
  geom_histogram(binwidth = 1, fill = "darkblue", color = "black") +
  scale_x_continuous(breaks = seq(min(sim_pop_sample_with_draws$.prediction), 
                                   max(sim_pop_sample_with_draws$.prediction), 
                                   by = 8)) +
  theme_minimal(base_size = 14) +
  labs(x = "RPP predicted CFT 20-R score", y = "Count") +
  facet_wrap(~ age_group, scales = "free_y")  # Create a separate facet for each level of the variable "age"
```






```{r}
sessionInfo()
```
