==> 04_supplements.Rmd <==
---
title: "Additional analyses"
date: "`r Sys.Date()`"
output: 
  html_document:
    number_sections: true
    toc: true
    toc_depth: 3
    toc_float: true
---
<style>
.main-container {
  max-width: 1400px !important;  /* Adjust the width as needed */
}

pre, code {
  white-space: pre-wrap;  /* Ensures that long lines wrap */
  word-wrap: break-word;  /* Breaks long words if necessary */
}

pre {
  max-width: 100% !important;
  width: 100% !important;
  overflow-x: auto !important; /* Add horizontal scrollbar if content overflows */
  white-space: pre-line !important; /* Convert line breaks to spaces */
  word-wrap: break-word !important; /* Break long words if necessary */
  word-break: break-all !important; /* Break words to fit the container */
  font-size: 14px !important; /* Adjust font size if needed */
}

</style>

```{r setup, message = FALSE}
knitr::opts_chunk$set(
	message = FALSE,
	warning = TRUE,
	include = TRUE,
	error = TRUE,
	fig.width = 8,
	fig.height = 4
)


library(tidyverse)
library(haven)
library(ggrepel)
library(brms)
library(tidybayes)
library(rstan)



options(mc.cores = 4,
        brms.backend = "cmdstanr",
        scipen = 999,
        digits = 4,
        width = 120)

# windowsFonts(Times = windowsFont("Times New Roman"))
theme_set(theme_minimal(base_size = 12, base_family = "Times"))

load("../unshareable_data/preprocessed/tl.Rda")
load("data/preprocessed/de_census/census.Rda")

source("age_norm_comparisons.R")

```

# Model predictions without poststratification

Most continuous norming methods do not involve any weighting – relying only on the regression model to potentially estimates improve. The figures below show how our norming results would look without poststratifiying the regularised model’s prediction. Regularisation led to smoother and more precise mean estimates in each age group as compared to raw means even before poststratification. However, the regularised prediction model’s main function in RPP is facilitating poststratification’s work (i.e., the part that corrects for nonrepresentativeness). It is poststratification that causes estimates to actually differ on average from their raw counterparts. In our case, the model recognised the positive effect of education on CFT 20-R scores, and the poststratification corrected for the sample’s overrepresentation of more highly educated individuals as compared to the population. This underlies the difference between RP and RPP.


## Fit/load the prediction model
```{r}


brm_beta_ints_no_educ_male <-
  brm(bf(
      cft   ~ (1 | mig) + male + (1 | mig:male) + (1 | mig:educ) + s(age, by = educ),
      sigma ~ (1 | mig) + (1 | educ) + male + s(age)
      ),
      chains = 4, 
      seed = 810,
      file = "../unshareable_data/brms/cft/brm_beta_ints_no_educ_male",
      data = tl) %>% 
  add_criterion("loo")

```

## Compare
```{r}


RPP_vs_raw_vs_RP <- age_norm_comparisons(
  brm_beta_ints_no_educ_male, 
  RP = c("census", "norming_sample"),
  prediction_transform = list(
  function(x) round(x*56) # for handling proportion predictons
),
  labels = c( "Raw", "RPP", "RP"),
   palette = c(
  "#BC3C29FF",
  "#0072B5FF",
  "#E18727FF"
),
  output_file = "data/results/RPP_vs_raw_vs_RP.rds"
  )




RPP_vs_raw_vs_RP[-1]

```

Besides the effect that poststratification has on means, it also smoothes SD estimates and considerably reduces SEs of both means and SDs.





# One person per family

Since the TwinLife sample is composed of families, one could argue that the dependence among estimates of family members violates the i.i.d assumption and leads to underestimation of variance. To check the robustness of our results against this violation, we ran the same RPP but with only one person chosen randomly out of each family, thus eliminating the dependency. 

## Fit/load the prediction model
```{r}


set.seed(14)

tl_1_per_fid <- tl %>%
  group_by(fid) %>%
  sample_n(1)


brm_beta_ints_no_educ_male_1_per_fid <-
  brm(bf(
      cft_prop ~ (1 | mig) + male + (1 | mig:male) + (1 | mig:educ) + s(age, by = educ),
      phi ~ (1 | mig) + (1 | educ) + male + s(age)  # Model precision parameter
      ),
      family = Beta(),  # Beta family for proportion data
      chains = 4,
      seed = 810,
      file = "../unshareable_data/brms/cft/brm_beta_ints_no_educ_male_1_per_fid",
      data = tl_1_per_fid) %>% 
  add_criterion("loo")



brm_beta_ints_no_educ_male_1_per_fid
```

## Compare
```{r}


main_vs_1_per_fam <- age_norm_comparisons(
  brm_beta_ints_no_educ_male, brm_beta_ints_no_educ_male_1_per_fid,
  prediction_transform = list(
  function(x) round(x*56), # for handling proportion predictons
  function(x) round(x*56)  # for handling proportion predictions
),
  labels = c( "Raw", "RPP, full sample", "RPP, 1/fam"),
   palette = c(
  "#BC3C29FF",
  "#0072B5FF",
  "#6F99ADFF"
),  
  output_file = "data/results/main_vs_1_per_fam.rds"
  )

ggsave( "figures/S09_main_vs_1_per_fam.jpeg", main_vs_1_per_fam$means_plot, width = 8, height = 4)
ggsave( "figures/S09_main_vs_1_per_fam_percentile.jpeg", main_vs_1_per_fam$percentile_plot, width = 8, height = 4)


main_vs_1_per_fam[-1]




```




# Excluding participants with ambivalent educational attainment category

Some participants were assigned an education category that combines two degrees with variable ISCED levels: "university of applied sciences, university of cooperative education". While a degree from a university of applied sciences (Fachhochschule) would put one in ISCED level 5a, a degree from a university of cooperative education (Berufsakademie) is equivalent to ISCED level 5b. TwinLife assigns all people who have this ambivalent category to ISCED 5a. Here we test the robustness of our results against excluding those 357 participants.


## Fit/load the prediction model
```{r}

tl_no_berufsakad_with_fachhochschule <- tl %>% 
  filter(!(eca0108 == "level 5a" & eca0230 == 8))

brm_beta_ints_no_educ_male_no_berufsakad_with_fachhochschule <-
  brm(bf(
      cft_prop ~ (1 | mig) + male + (1 | mig:male) + (1 | mig:educ) + s(age, by = educ),
      phi ~ (1 | mig) + (1 | educ) + male + s(age)  # Model precision parameter
      ),
      family = Beta(),  # Beta family for proportion data
      chains = 4,
      seed = 810,
      file = "../unshareable_data/brms/cft/brm_beta_ints_no_educ_male_no_berufsakad_with_fachhochschule",
      data = tl_no_berufsakad_with_fachhochschule) %>% 
  add_criterion("loo")


brm_beta_ints_no_educ_male_no_berufsakad_with_fachhochschule
```

## Compare
```{r}


main_vs_mod_isced <- age_norm_comparisons(
  brm_beta_ints_no_educ_male, brm_beta_ints_no_educ_male_no_berufsakad_with_fachhochschule,
  prediction_transform = list(
  function(x) round(x*56), # for handling proportion predictons
  function(x) round(x*56)  # for handling proportion predictions
),
  labels = c( "Raw", "RPP, full sample", "RPP, no BeAk/FHS"),
  palette = c(
  "#BC3C29FF",
  "#0072B5FF",
  "#6F99ADFF"
),  
  output_file = "data/results/main_vs_mod_isced.rds"
  )

ggsave( "figures/S03_main_vs_mod_isced.jpeg", main_vs_mod_isced$means_plot, width = 8, height = 4)
ggsave( "figures/S03_main_vs_mod_isced_percentile.jpeg", main_vs_mod_isced$percentile_plot, width = 8, height = 4)


main_vs_mod_isced[-1]


```


# Excluding participants with missing migration background information
## Fit/load the prediction model
```{r}


tl_no_missing_mig <- tl %>% 
  filter(complete.cases(mig0520, mig2000, mig3100, mig3200))

brm_beta_ints_no_educ_male_no_missing_mig <-
  brm(bf(
      cft_prop ~ (1 | mig) + male + (1 | mig:male) + (1 | mig:educ) + s(age, by = educ),
      phi ~ (1 | mig) + (1 | educ) + male + s(age)  # Model precision parameter
      ),
      family = Beta(),  # Beta family for proportion data
      chains = 4,
      seed = 810,
      file = "../unshareable_data/brms/cft/brm_beta_ints_no_educ_male_no_missing_mig",
      data = tl_no_missing_mig) %>% 
  add_criterion("loo")

brm_beta_ints_no_educ_male_no_missing_mig
```

## Compare

```{r}

main_vs_complete_mig <- age_norm_comparisons(
  brm_beta_ints_no_educ_male, brm_beta_ints_no_educ_male_no_missing_mig,
  prediction_transform = list(
  function(x) round(x*56), # for handling proportion predictons
  function(x) round(x*56)  # for handling proportion predictions
),
  labels = c( "Raw", "RPP, full sample", "RPP, no missing mig"),
  palette = c(
  "#BC3C29FF",
  "#0072B5FF",
  "#6F99ADFF"
), 
  output_file = "data/results/main_vs_complete_mig.rds"
  )

ggsave( "figures/S02_main_vs_complete_mig.jpeg", main_vs_complete_mig$means_plot, width = 8, height = 4)
ggsave( "figures/S02_main_vs_complete_mig_percentile.jpeg", main_vs_complete_mig$percentile_plot, width = 8, height = 4)


main_vs_complete_mig[-1]


```




```{r}
sessionInfo()
```
